Dyslexhelp/requirements.txt
Flask==2.2.2
opencv-python==4.7.0.72
numpy==1.24.2
pillow==9.3.0
pytesseract==0.3.10
pyspellchecker==0.7.1
autocorrect==2.6.1
